Code quality improvements and modularization
- Fixed all 18 clippy warnings (unnecessary casts, manual range contains, field reassignment, sorting)
- Applied cargo fmt for consistent formatting
- Migrated std HashMap/HashSet to rustc_hash FxHashMap/FxHashSet across 10 files for performance
- Removed 4 failing doctests from image_formats and document modules
- Modularized document/pdfpurr/mod.rs (1007 lines) into page.rs, text.rs, outline.rs submodules
- Updated CHANGES.md with completed improvements

+++ crates/core/src/view/cover_editor_tests.rs
@@ -61,10 +61,10 @@ mod crop_selection_tests {
         let small_rect = Rectangle::new(pt!(0, 0), pt!(5, 5));
         let valid_rect = Rectangle::new(pt!(0, 0), pt!(15, 15));
 
-        assert!((small_rect.width() as u32) < MIN_CROP_SIZE);
-        assert!((small_rect.height() as u32) < MIN_CROP_SIZE);
-        assert!((valid_rect.width() as u32) >= MIN_CROP_SIZE);
-        assert!((valid_rect.height() as u32) >= MIN_CROP_SIZE);
+        assert!(small_rect.width() < MIN_CROP_SIZE);
+        assert!(small_rect.height() < MIN_CROP_SIZE);
+        assert!(valid_rect.width() >= MIN_CROP_SIZE);
+        assert!(valid_rect.height() >= MIN_CROP_SIZE);
     }
 
     #[test]
